Create problem no. 199
Binary Tree Right Side View Solution

 Overview:

This repository contains my solution to the **#LeetCode** problem **"Binary Tree Right Side View."** The goal of this problem is to return the values of the nodes you can see when standing on the right side of a binary tree, ordered from top to bottom.


Problem Statement:

Given the root of a binary tree, imagine yourself standing on the right side of it, and return the values of the nodes you can see ordered from top to bottom.


Example:

1. **Input**: `[1,2,3,null,5,null,4]`  
   **Output**: `[1,3,4]`

2. **Input**: `[1,null,3]`  
   **Output**: `[1,3]`

3. **Input**: `[]`  
   **Output**: `[]`



Approach:

- The main method **#rightSideView** is implemented using the breadth-first search (BFS) technique to traverse the binary tree.

- The last node at each level is identified and added to the result list.


Complexity Analysis:

- **Time Complexity**: \(O(N)\)  
  (Where N is the total number of nodes in the tree, as we visit each node once.)

- **Space Complexity**: \(O(W)\)  
  (Where W is the maximum width of the tree, which is stored in the queue.)


Technologies Used:

- Java


How to Run:

1. Clone this code to your local machine.
2. Set up a Java development environment.
3. Run the code to see the solution to the problem.


Conclusion

This solution is part of my **100 Days of Code** challenge, where I aim to solve daily DSA problems. Feel free to follow me on this journey!