Create problem no. 240
Problem 240: Search a 2D Matrix II

Description

You are given an m x n integer matrix matrix that is sorted in ascending order both row-wise and column-wise. You need to write a function that searches for a target value in the matrix. If the target exists, return true; otherwise, return false.

Approach:

Start the search from the bottom-left corner of the matrix.

If the current element is equal to the target, return true.

If the current element is greater than the target, move one row up.

If the current element is less than the target, move one column to the right.

Continue this process until you either find the target or exit the boundaries of the matrix.

If the loop ends without finding the target, return false.

Complexity:

Time Complexity:-  O(m+n)

Space Complexity:- O(1)